+++ app/controllers/cloudflare/email/management/styles_controller.rb
@@ -1,12 +1,8 @@
+require File.join(MailboxKit::ROOT, "app/controllers/mailbox_kit/management/styles_controller.rb")
 module Cloudflare
   module Email
     module Management
-      class StylesController < ActionController::Base
-        def show
-          response.headers["X-Content-Type-Options"] = "nosniff"
-          send_data File.binread(File.expand_path("../../../../../lib/cloudflare/email/management/management.css", __dir__)),
-            type: "text/css; charset=utf-8", disposition: "inline"
-        end
+      class StylesController < MailboxKit::Management::StylesController
       end
     end
   end