+++ mailbox-kit/lib/mailbox_kit/management.rb
@@ -1,0 +1,6 @@
+require "rails/engine"
+require "mailbox-kit"
+require "mailbox_kit/railtie"
+require "mailbox_kit/management/adapter"
+require "mailbox_kit/management/configuration"
+require "mailbox_kit/management/engine"