+++ mailbox-kit/lib/mailbox_kit/management/configuration.rb
@@ -1,0 +1,15 @@
+module MailboxKit
+  module Management
+    class Configuration
+      attr_accessor :adapter, :back_path
+    end
+
+    def self.configuration
+      @configuration ||= Configuration.new
+    end
+
+    def self.configure
+      yield configuration
+    end
+  end
+end